browser: Move gm and sharp usages to separate file.
