+++ lib/index.ts
@@ -20,5 +20,3 @@
 export { Table } from './vpt/table';
 export { OleCompoundDoc, Storage } from './io/ole-doc';
 export { Logger } from './util/logger';
-// export { NodeBinaryReader } from './io/binary-reader.node';
-// export { BrowserBinaryReader } from './io/binary-reader.browser';
