+++ lib/refs.browser.ts
@@ -1,2 +1,3 @@
 
 export { BrowserBinaryReader as BinaryReader } from './io/binary-reader.browser';
+export { loadImage, getRawImage, streamImage, BrowserImage as Image } from './gltf/image.browser';
