+++ lib/refs.node.ts
@@ -1,2 +1,3 @@
 
 export { NodeBinaryReader as BinaryReader } from './io/binary-reader.node';
+export { loadImage, getRawImage, streamImage, NodeImage as Image } from './gltf/image.node';
